rearrange functions in pkg/api/validate, should be a no-op
