+++ pkg/api/validate/validators.go
@@ -3,87 +3,9 @@ package validate
 import (
 	"fmt"
 	"net"
-	"strconv"
-	"strings"
-
-	"github.com/google/uuid"
 
 	"github.com/openshift/openshift-azure/pkg/api"
 )
-
-func init() {
-	var err error
-
-	// TODO: we probably need to bite the bullet and make these configurable.
-	_, clusterNetworkCIDR, err = net.ParseCIDR("10.128.0.0/14")
-	if err != nil {
-		panic(err)
-	}
-
-	_, serviceNetworkCIDR, err = net.ParseCIDR("172.30.0.0/16")
-	if err != nil {
-		panic(err)
-	}
-}
-
-var validAgentPoolProfileRoles = map[api.AgentPoolProfileRole]struct{}{
-	api.AgentPoolProfileRoleCompute: {},
-	api.AgentPoolProfileRoleInfra:   {},
-	api.AgentPoolProfileRoleMaster:  {},
-}
-
-var validRouterProfileNames = map[string]struct{}{
-	"default": {},
-}
-
-func isValidHostname(h string) bool {
-	return len(h) <= 255 && rxRfc1123.MatchString(h)
-}
-
-func isValidCloudAppHostname(h, location string) bool {
-	if !rxCloudDomainLabel.MatchString(h) {
-		return false
-	}
-	return strings.HasSuffix(h, "."+location+".cloudapp.azure.com") && strings.Count(h, ".") == 4
-}
-
-func isValidIPV4CIDR(cidr string) bool {
-	ip, net, err := net.ParseCIDR(cidr)
-	if err != nil {
-		return false
-	}
-	if ip.To4() == nil {
-		return false
-	}
-	if net == nil || !ip.Equal(net.IP) {
-		return false
-	}
-	return true
-}
-
-func IsValidBlobContainerName(c string) bool {
-	// https://docs.microsoft.com/en-us/rest/api/storageservices/naming-and-referencing-containers--blobs--and-metadata
-	if !rxBlobContainerName.MatchString(c) {
-		return false
-	}
-
-	if strings.Trim(c, "-") != c || strings.Contains(c, "--") {
-		return false
-	}
-
-	return true
-}
-
-func vnetContainsSubnet(vnet, subnet *net.IPNet) bool {
-	vnetbits, _ := vnet.Mask.Size()
-	subnetbits, _ := subnet.Mask.Size()
-	if vnetbits > subnetbits {
-		// e.g., vnet is a /24, subnet is a /16: vnet cannot contain subnet.
-		return false
-	}
-
-	return vnet.IP.Equal(subnet.IP.Mask(vnet.Mask))
-}
 
 func validateContainerService(c *api.OpenShiftManagedCluster, externalOnly bool) (errs []error) {
 	if c == nil {
@@ -132,42 +54,81 @@ func validateProperties(p *api.Properties, location string, externalOnly bool) (
 	return
 }
 
-func validateAuthProfile(ap *api.AuthProfile) (errs []error) {
-	if ap == nil {
-		errs = append(errs, fmt.Errorf("properties.authProfile cannot be nil"))
+func validateProvisioningState(ps api.ProvisioningState) (errs []error) {
+	switch ps {
+	case "",
+		api.Creating,
+		api.Updating,
+		api.AdminUpdating,
+		api.Failed,
+		api.Succeeded,
+		api.Deleting,
+		api.Migrating,
+		api.Upgrading:
+	default:
+		errs = append(errs, fmt.Errorf("invalid properties.provisioningState %q", ps))
+	}
+	return
+}
+
+func validateNetworkProfile(np *api.NetworkProfile) (errs []error) {
+	if np == nil {
+		errs = append(errs, fmt.Errorf("networkProfile cannot be nil"))
 		return
 	}
-
-	if len(ap.IdentityProviders) != 1 {
-		errs = append(errs, fmt.Errorf("invalid properties.authProfile.identityProviders length"))
+	if np.VnetID != "" && !rxVNetID.MatchString(np.VnetID) {
+		errs = append(errs, fmt.Errorf("invalid properties.networkProfile.vnetId %q", np.VnetID))
 	}
-	//check supported identity providers
-	for _, ip := range ap.IdentityProviders {
-		switch provider := ip.Provider.(type) {
-		case (*api.AADIdentityProvider):
-			if ip.Name != "Azure AD" {
-				errs = append(errs, fmt.Errorf("invalid properties.authProfile.identityProviders name"))
-			}
-			if provider.Secret == "" {
-				errs = append(errs, fmt.Errorf("invalid properties.authProfile.AADIdentityProvider secret %q", provider.Secret))
-			}
-			if provider.ClientID == "" {
-				errs = append(errs, fmt.Errorf("invalid properties.authProfile.AADIdentityProvider clientId %q", provider.ClientID))
-			}
-			if provider.TenantID == "" {
-				errs = append(errs, fmt.Errorf("invalid properties.authProfile.AADIdentityProvider tenantId %q", provider.TenantID))
-			}
-			if provider.CustomerAdminGroupID != nil && !isValidUUID(*provider.CustomerAdminGroupID) {
-				errs = append(errs, fmt.Errorf("invalid properties.authProfile.AADIdentityProvider customerAdminGroupId %q", *provider.CustomerAdminGroupID))
+	if !isValidIPV4CIDR(np.VnetCIDR) {
+		errs = append(errs, fmt.Errorf("invalid properties.networkProfile.vnetCidr %q", np.VnetCIDR))
+	}
+	if np.PeerVnetID != nil && !rxVNetID.MatchString(*np.PeerVnetID) {
+		errs = append(errs, fmt.Errorf("invalid properties.networkProfile.peerVnetId %q", *np.PeerVnetID))
+	}
+	return
+}
+
+func validateRouterProfiles(rps []api.RouterProfile, location string, externalOnly bool) (errs []error) {
+	rpmap := map[string]api.RouterProfile{}
+
+	for _, rp := range rps {
+		if _, found := validRouterProfileNames[rp.Name]; !found {
+			errs = append(errs, fmt.Errorf("invalid properties.routerProfiles[%q]", rp.Name))
+		}
+
+		if _, found := rpmap[rp.Name]; found {
+			errs = append(errs, fmt.Errorf("duplicate properties.routerProfiles %q", rp.Name))
+		}
+		rpmap[rp.Name] = rp
+
+		errs = append(errs, validateRouterProfile(rp, location, externalOnly)...)
+	}
+	if !externalOnly {
+		for name := range validRouterProfileNames {
+			if _, found := rpmap[name]; !found {
+				errs = append(errs, fmt.Errorf("invalid properties.routerProfiles[%q]", name))
 			}
 		}
 	}
 	return
 }
 
-func isValidUUID(u string) bool {
-	_, err := uuid.Parse(u)
-	return err == nil
+func validateRouterProfile(rp api.RouterProfile, location string, externalOnly bool) (errs []error) {
+	if rp.Name == "" {
+		errs = append(errs, fmt.Errorf("invalid properties.routerProfiles[%q].name %q", rp.Name, rp.Name))
+	}
+
+	// TODO: consider ensuring that PublicSubdomain is of the form
+	// <string>.<location>.azmosa.io
+	if rp.PublicSubdomain != "" && !isValidHostname(rp.PublicSubdomain) {
+		errs = append(errs, fmt.Errorf("invalid properties.routerProfiles[%q].publicSubdomain %q", rp.Name, rp.PublicSubdomain))
+	}
+
+	if !externalOnly && rp.FQDN != "" && !isValidCloudAppHostname(rp.FQDN, location) {
+		errs = append(errs, fmt.Errorf("invalid properties.routerProfiles[%q].fqdn %q", rp.Name, rp.FQDN))
+	}
+
+	return
 }
 
 func validateAgentPoolProfiles(apps []api.AgentPoolProfile, vnet *net.IPNet) (errs []error) {
@@ -263,40 +224,37 @@ func validateAgentPoolProfile(app api.AgentPoolProfile, vnet *net.IPNet) (errs [
 	return
 }
 
-func IsValidAgentPoolHostname(hostname string) bool {
-	parts := strings.Split(hostname, "-")
-	switch len(parts) {
-	case 2: // master-XXXXXX
-		if parts[0] != "master" ||
-			len(parts[1]) != 6 {
-			return false
-		}
-		_, err := strconv.ParseUint(parts[1], 36, 64)
-		if err != nil {
-			return false
-		}
-
-	case 3: // something-XXXXXXXXXX-XXXXXX
-		if !rxAgentPoolProfileName.MatchString(parts[0]) ||
-			parts[0] == "master" ||
-			len(parts[1]) != 10 ||
-			len(parts[2]) != 6 {
-			return false
-		}
-		_, err := strconv.ParseUint(parts[1], 10, 64)
-		if err != nil {
-			return false
-		}
-		_, err = strconv.ParseUint(parts[2], 36, 64)
-		if err != nil {
-			return false
-		}
-
-	default:
-		return false
+func validateAuthProfile(ap *api.AuthProfile) (errs []error) {
+	if ap == nil {
+		errs = append(errs, fmt.Errorf("properties.authProfile cannot be nil"))
+		return
 	}
 
-	return true
+	if len(ap.IdentityProviders) != 1 {
+		errs = append(errs, fmt.Errorf("invalid properties.authProfile.identityProviders length"))
+	}
+	//check supported identity providers
+	for _, ip := range ap.IdentityProviders {
+		switch provider := ip.Provider.(type) {
+		case (*api.AADIdentityProvider):
+			if ip.Name != "Azure AD" {
+				errs = append(errs, fmt.Errorf("invalid properties.authProfile.identityProviders name"))
+			}
+			if provider.Secret == "" {
+				errs = append(errs, fmt.Errorf("invalid properties.authProfile.AADIdentityProvider secret %q", provider.Secret))
+			}
+			if provider.ClientID == "" {
+				errs = append(errs, fmt.Errorf("invalid properties.authProfile.AADIdentityProvider clientId %q", provider.ClientID))
+			}
+			if provider.TenantID == "" {
+				errs = append(errs, fmt.Errorf("invalid properties.authProfile.AADIdentityProvider tenantId %q", provider.TenantID))
+			}
+			if provider.CustomerAdminGroupID != nil && !isValidUUID(*provider.CustomerAdminGroupID) {
+				errs = append(errs, fmt.Errorf("invalid properties.authProfile.AADIdentityProvider customerAdminGroupId %q", *provider.CustomerAdminGroupID))
+			}
+		}
+	}
+	return
 }
 
 func validateFQDN(p *api.Properties, location string) (errs []error) {
@@ -306,83 +264,6 @@ func validateFQDN(p *api.Properties, location string) (errs []error) {
 	}
 	if p.FQDN == "" || !isValidCloudAppHostname(p.FQDN, location) {
 		errs = append(errs, fmt.Errorf("invalid properties.fqdn %q", p.FQDN))
-	}
-	return
-}
-
-func validateNetworkProfile(np *api.NetworkProfile) (errs []error) {
-	if np == nil {
-		errs = append(errs, fmt.Errorf("networkProfile cannot be nil"))
-		return
-	}
-	if np.VnetID != "" && !rxVNetID.MatchString(np.VnetID) {
-		errs = append(errs, fmt.Errorf("invalid properties.networkProfile.vnetId %q", np.VnetID))
-	}
-	if !isValidIPV4CIDR(np.VnetCIDR) {
-		errs = append(errs, fmt.Errorf("invalid properties.networkProfile.vnetCidr %q", np.VnetCIDR))
-	}
-	if np.PeerVnetID != nil && !rxVNetID.MatchString(*np.PeerVnetID) {
-		errs = append(errs, fmt.Errorf("invalid properties.networkProfile.peerVnetId %q", *np.PeerVnetID))
-	}
-	return
-}
-
-func validateRouterProfiles(rps []api.RouterProfile, location string, externalOnly bool) (errs []error) {
-	rpmap := map[string]api.RouterProfile{}
-
-	for _, rp := range rps {
-		if _, found := validRouterProfileNames[rp.Name]; !found {
-			errs = append(errs, fmt.Errorf("invalid properties.routerProfiles[%q]", rp.Name))
-		}
-
-		if _, found := rpmap[rp.Name]; found {
-			errs = append(errs, fmt.Errorf("duplicate properties.routerProfiles %q", rp.Name))
-		}
-		rpmap[rp.Name] = rp
-
-		errs = append(errs, validateRouterProfile(rp, location, externalOnly)...)
-	}
-	if !externalOnly {
-		for name := range validRouterProfileNames {
-			if _, found := rpmap[name]; !found {
-				errs = append(errs, fmt.Errorf("invalid properties.routerProfiles[%q]", name))
-			}
-		}
-	}
-	return
-}
-
-func validateRouterProfile(rp api.RouterProfile, location string, externalOnly bool) (errs []error) {
-	if rp.Name == "" {
-		errs = append(errs, fmt.Errorf("invalid properties.routerProfiles[%q].name %q", rp.Name, rp.Name))
-	}
-
-	// TODO: consider ensuring that PublicSubdomain is of the form
-	// <string>.<location>.azmosa.io
-	if rp.PublicSubdomain != "" && !isValidHostname(rp.PublicSubdomain) {
-		errs = append(errs, fmt.Errorf("invalid properties.routerProfiles[%q].publicSubdomain %q", rp.Name, rp.PublicSubdomain))
-	}
-
-	if !externalOnly && rp.FQDN != "" && !isValidCloudAppHostname(rp.FQDN, location) {
-		errs = append(errs, fmt.Errorf("invalid properties.routerProfiles[%q].fqdn %q", rp.Name, rp.FQDN))
-	}
-
-	return
-}
-
-func validateProvisioningState(ps api.ProvisioningState) (errs []error) {
-	switch ps {
-	case "",
-		api.Creating,
-		api.Updating,
-		api.AdminUpdating,
-		api.Failed,
-		api.Succeeded,
-		api.Deleting,
-		api.Migrating,
-		api.Upgrading:
-	default:
-		errs = append(errs, fmt.Errorf("invalid properties.provisioningState %q", ps))
 	}
 	return
 }
@@ -403,22 +284,4 @@ func validateVMSize(app api.AgentPoolProfile, runningUnderTest bool) (errs []err
 		}
 	}
 	return errs
-}
-
-func isValidMasterAndInfraVMSize(size api.VMSize, runningUnderTest bool) bool {
-	if runningUnderTest && size == api.StandardD2sV3 {
-		return true
-	}
-
-	_, found := validMasterAndInfraVMSizes[size]
-	return found
-}
-
-func isValidComputeVMSize(size api.VMSize, runningUnderTest bool) bool {
-	if runningUnderTest && size == api.StandardD2sV3 {
-		return true
-	}
-
-	_, found := validComputeVMSizes[size]
-	return found
 }
